backend/sheets_writer.py
import google.auth
from googleapiclient.discovery import build
from config import config, logger

class SheetsWriter:
    """
    Modular integration with Google Sheets for data persistence.
    """
    def __init__(self, sheet_id, credentials_file):
        self.sheet_id = sheet_id
        self.credentials_file = credentials_file
        # self.service = build('sheets', 'v4', credentials=credentials)
        logger.info("Sheets Writer initialized for Sheet ID: %s", self.sheet_id)

    def append_job_data(self, data_row):
        """
        Appends a row of processed job data to the linked Google Sheet.
        """
        logger.info("Appending job data: %s", data_row)
        # TODO: Implement Sheets API write logic
        return True

    def get_existing_data(self):
        """
        Fetches current Google Sheets data for checking duplicates.
        """
        logger.info("Fetching existing Google Sheets data.")
        # TODO: Implement Sheets API read logic
        return []

# Create a global instance
# sheets_writer = SheetsWriter(config.GOOGLE_SHEET_ID, config.GOOGLE_CREDENTIALS_FILE)
